Moved helper class to Helper folder.
