+++ Tests/DICOM.Tests.csproj
@@ -111,6 +111,7 @@
     <Compile Include="DicomFileMetaInformationTest.cs" />
     <Compile Include="DicomParseableTest.cs" />
     <Compile Include="Helpers\NLogHelper.cs" />
+    <Compile Include="Helpers\TestOutputHelperLogger.cs" />
     <Compile Include="Imaging\Codec\DicomCodecExtensionsTest.cs" />
     <Compile Include="Imaging\ColorTableTest.cs" />
     <Compile Include="Imaging\DicomImageTest.cs" />
@@ -161,7 +162,6 @@
     <Compile Include="Printing\FilmBoxTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Serialization\JsonDicomConverterTest.cs" />
-    <Compile Include="TestOutputHelperLogger.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\DICOM.Native\x64\DICOM.Native64.vcxproj">
