+++ Tests/Helpers/TestOutputHelperLogger.cs
@@ -1,26 +1,25 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Dicom.Log;
-using Xunit.Abstractions;
+﻿// Copyright (c) 2012-2016 fo-dicom contributors.
+// Licensed under the Microsoft Public License (MS-PL).
 
-namespace Dicom
+namespace Dicom.Helpers
 {
+    using Dicom.Log;
+
+    using Xunit.Abstractions;
+
     internal class TestOutputHelperLogger: Logger
     {
         private readonly ITestOutputHelper _testOutputHelper;
         
         public TestOutputHelperLogger(ITestOutputHelper testOutputHelper)
         {
-            this._testOutputHelper = testOutputHelper;
+            _testOutputHelper = testOutputHelper;
         }
 
         public override void Log(LogLevel level, string msg, params object[] args)
         {
             var positionMsg = NameFormatToPositionalFormat(msg);
-            this._testOutputHelper.WriteLine(level.ToString() + ": " + positionMsg, args);
+            _testOutputHelper.WriteLine(level + ": " + positionMsg, args);
         }
     }
 }
