+++ Tests/Network/DicomClientTest.cs
@@ -1,19 +1,19 @@
 ﻿// Copyright (c) 2012-2016 fo-dicom contributors.
 // Licensed under the Microsoft Public License (MS-PL).
 
-using System.Linq;
-using Xunit.Abstractions;
-
 namespace Dicom.Network
 {
     using System;
+    using System.Linq;
     using System.Text;
     using System.Threading;
     using System.Threading.Tasks;
 
+    using Dicom.Helpers;
     using Dicom.Log;
 
     using Xunit;
+    using Xunit.Abstractions;
 
     [Collection("Network"), Trait("Category", "Network")]
     public class DicomClientTest
@@ -147,18 +147,15 @@ namespace Dicom.Network
         [Theory]
         [InlineData(20)]
         [InlineData(200)]
-        //[InlineData(1000)]
         public async Task SendAsync_MultipleTimes_AllRecognized(int expected)
         {
             int port = Ports.GetNext();
-            
-            using (var server = new DicomServer<DicomCEchoProvider>(
-                port, 
-                options: new DicomServiceOptions()
-                {
-                    //LogDataPDUs = true, LogDimseDatasets = true
-                }, 
-                logger: new TestOutputHelperLogger(_testOutputHelper)))
+
+            using (
+                var server = new DicomServer<DicomCEchoProvider>(
+                    port,
+                    options: new DicomServiceOptions(),
+                    logger: new TestOutputHelperLogger(_testOutputHelper)))
             {
                 await Task.Delay(500);
                 Assert.True(server.IsListening, "Server is not listening");
@@ -168,22 +165,20 @@ namespace Dicom.Network
                 var client = new DicomClient();
                 for (var i = 0; i < expected; i++)
                 {
-                    client.AddRequest(new DicomCEchoRequest()
-                    {
-                        OnResponseReceived = (req, res) =>
-                        {
-                            this._testOutputHelper.WriteLine("Response #{0}", i);
-                            Interlocked.Add(ref actual, 1);
-                        }
-                    });
-                    this._testOutputHelper.WriteLine("Sending #{0}", i);
-                    await Task.WhenAny(
-                        client.SendAsync("127.0.0.1", port, false, "SCU", "ANY-SCP"),
-                        Task.Delay(1000)
-                        );
-                    this._testOutputHelper.WriteLine("Sent (or timed out) #{0}", i);
+                    client.AddRequest(
+                        new DicomCEchoRequest
+                            {
+                                OnResponseReceived = (req, res) =>
+                                    {
+                                        _testOutputHelper.WriteLine("Response #{0}", i);
+                                        Interlocked.Add(ref actual, 1);
+                                    }
+                            });
+                    _testOutputHelper.WriteLine("Sending #{0}", i);
+                    await Task.WhenAny(client.SendAsync("127.0.0.1", port, false, "SCU", "ANY-SCP"), Task.Delay(1000));
+                    _testOutputHelper.WriteLine("Sent (or timed out) #{0}", i);
                 }
-                
+
                 Assert.Equal(expected, actual);
             }
         }
@@ -194,39 +189,39 @@ namespace Dicom.Network
         public async Task SendAsync_MultipleTimesParallel_AllRecognized(int expected)
         {
             int port = Ports.GetNext();
-            
-            using (var server = new DicomServer<DicomCEchoProvider>(
-                port,
-                options: new DicomServiceOptions()
-                {
-                    //LogDataPDUs = true, LogDimseDatasets = true
-                },
-                logger: new TestOutputHelperLogger(_testOutputHelper)))
+
+            using (
+                var server = new DicomServer<DicomCEchoProvider>(
+                    port,
+                    options: new DicomServiceOptions(),
+                    logger: new TestOutputHelperLogger(_testOutputHelper)))
             {
                 await Task.Delay(500);
                 Assert.True(server.IsListening, "Server is not listening");
 
                 var actual = 0;
 
-                var requests = Enumerable.Range(0, expected).Select(async requestIndex =>
-                {
-                    var client = new DicomClient();
-                    client.AddRequest(new DicomCEchoRequest()
-                    {
-                        OnResponseReceived = (req, res) =>
+                var requests = Enumerable.Range(0, expected).Select(
+                    async requestIndex =>
                         {
-                            this._testOutputHelper.WriteLine("Response #{0}", requestIndex);
-                            Interlocked.Add(ref actual, 1);
-                        }
-                    });
+                            var client = new DicomClient();
+                            client.AddRequest(
+                                new DicomCEchoRequest
+                                    {
+                                        OnResponseReceived = (req, res) =>
+                                            {
+                                                _testOutputHelper.WriteLine("Response #{0}", requestIndex);
+                                                Interlocked.Add(ref actual, 1);
+                                            }
+                                    });
 
-                    this._testOutputHelper.WriteLine("Sending #{0}", requestIndex);
-                    var sendTask = client.SendAsync("127.0.0.1", port, false, "SCU", "ANY-SCP");
-                    await Task.WhenAny(sendTask, Task.Delay(1000));
-                    this._testOutputHelper.WriteLine("Sent (or timed out) #{0}", requestIndex);
-                }).ToList();
+                            _testOutputHelper.WriteLine("Sending #{0}", requestIndex);
+                            var sendTask = client.SendAsync("127.0.0.1", port, false, "SCU", "ANY-SCP");
+                            await Task.WhenAny(sendTask, Task.Delay(1000));
+                            _testOutputHelper.WriteLine("Sent (or timed out) #{0}", requestIndex);
+                        }).ToList();
                 await Task.WhenAll(requests);
-                
+
 
                 Assert.Equal(expected, actual);
             }
